Replace `DownloadTasks` with composeable `DownloadTask`.
Massive cleanup to most of the download logic. Still a bit of duplication for in-memory downloads though, but thats a later cleanup.

The main goal here was allowing the outputs of downloads to be mutated, (pretty printing json files, etc), which required implementing an overarching 'task cache' result, stored in a companion file. This does mean we now validate all files are correct on disk.